% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YACT-package.R
\docType{package}
\name{YACT-package}
\alias{YACT}
\alias{YACT-package}
\title{YACT: Yet Anothr Cleaning Template for IMPACT Initiatives}
\description{
Provides an end-to-end pipeline for processing survey data, specifically designed for data and analysis staff within IMPACT Initiatives. This package utilizes the 'targets' framework to orchestrate and integrate various data cleaning, composition, and analysis tools that are currently available across different R packages developed by IMPACT. Its aim is to streamline and standardize the workflow from raw survey data to analytical outputs.
}
\author{
\strong{Maintainer}: Iyed Ghedamsi \email{iyed.ghedamsi@impact-initiatives.org}

Other contributors:
\itemize{
  \item IMPACT Initiatives [copyright holder]
}

}
\keyword{internal}
